refactor(casbin): replace the module with a Casbin v3-only major (in place)
Switches the existing v3/casbin module to Casbin v3 only and bumps its
path to github.com/gofiber/contrib/v3/casbin/v2 (major-branch layout:
the code stays in v3/casbin, the module path carries the /v2 suffix).

Casbin v2 is no longer supported here; existing users can pin the
previous major (github.com/gofiber/contrib/v3/casbin) at its last v1
tag. The middleware API (New/Config/RequiresPermissions/RoutePermission/
RequiresRoles) is unchanged; pass a Casbin v3 enforcer/adapter.

- drop the separate v2/ subdirectory and the dual-version approach
- README documents Casbin v3 usage with gorm-adapter/v3
- revert the test workflow split (single module again)
