--- .github/workflows/test-casbin-v2.yml
@@ -1,51 +1,0 @@
-name: "Test casbin v2 (Casbin v3)"
-
-on:
-  push:
-    branches:
-      - master
-      - main
-    paths:
-      - 'v3/casbin/v2/**/*.go'
-      - 'v3/casbin/v2/go.mod'
-      - 'v3/casbin/v2/go.sum'
-  pull_request:
-    paths:
-      - 'v3/casbin/v2/**/*.go'
-      - 'v3/casbin/v2/go.mod'
-      - 'v3/casbin/v2/go.sum'
-
-  workflow_dispatch:
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref || github.ref }}
-  cancel-in-progress: true
-
-permissions:
-  contents: read
-
-jobs:
-  Tests:
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    env:
-      GOWORK: off
-    strategy:
-      matrix:
-        go-version:
-          - 1.25.x
-    steps:
-      - name: Fetch Repository
-        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
-      - name: Install Go
-        uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6
-        with:
-          go-version: '${{ matrix.go-version }}'
-          cache-dependency-path: v3/casbin/v2/go.sum
-      - name: Run Test
-        uses: gofiber/.github/.github/actions/gotestsum@main
-        with:
-          working-directory: ./v3/casbin/v2
-          packages: ./...
-          rerun-fails: '2'
-          args: -race -count=1
