+++ v3/casbin/README.md
@@ -12,10 +12,11 @@ Casbin middleware for Fiber.
 
 **Compatible with Fiber v3.**
 
-> **Deprecated:** this module targets Casbin v2, which is frozen upstream. New
-> projects should use [`github.com/gofiber/contrib/v3/casbin/v2`](./v2),
-> which targets Casbin v3. Migrating from Casbin v2 to v3 is mostly an
-> import-path change; see the [Casbin upgrade guide](https://casbin.org/docs/upgrade).
+> This middleware targets **Casbin v3**. Casbin v2 is no longer supported here;
+> if you still need it, pin the previous major
+> (`github.com/gofiber/contrib/v3/casbin` at its last `v1` tag). Migrating from
+> Casbin v2 to v3 is mostly an import-path change; see the
+> [Casbin upgrade guide](https://casbin.org/docs/upgrade).
 
 ## Go version support
 
@@ -24,11 +25,11 @@ We only support the latest two versions of Go. Visit [https://go.dev/doc/devel/r
 ## Install
 ```sh
 go get -u github.com/gofiber/fiber/v3
-go get -u github.com/gofiber/contrib/v3/casbin
+go get -u github.com/gofiber/contrib/v3/casbin/v2
 ```
 choose an adapter from [here](https://casbin.org/docs/adapters)
 ```sh
-go get -u github.com/casbin/xorm-adapter
+go get -u github.com/casbin/gorm-adapter/v3
 ```
 
 ## Signature
@@ -42,7 +43,7 @@ casbin.New(config ...casbin.Config) *casbin.Middleware
 |:--------------|:--------------------------|:-----------------------------------------|:--------------------------------------------------------------|
 | ModelFilePath | `string`                  | Model file path                        | `"./model.conf"`                                                    |
 | PolicyAdapter | `persist.Adapter`         | Database adapter for policies            | `./policy.csv`                                                      |
-| Enforcer      | `*casbin.Enforcer`        | Custom casbin enforcer                   | `Middleware generated enforcer using ModelFilePath & PolicyAdapter` |
+| Enforcer      | `*casbin.Enforcer`        | Custom casbin v3 enforcer                | `Middleware generated enforcer using ModelFilePath & PolicyAdapter` |
 | Lookup        | `func(fiber.Ctx) string`  | Look up for current subject              | `""`                                                              |
 | Unauthorized  | `func(fiber.Ctx) error`   | Response body for unauthorized responses | `Unauthorized`                                                      |
 | Forbidden     | `func(fiber.Ctx) error`   | Response body for forbidden responses    | `Forbidden`                                                         |
@@ -58,19 +59,20 @@ package main
 
 import (
   "github.com/gofiber/fiber/v3"
-  "github.com/gofiber/contrib/v3/casbin"
+  "github.com/gofiber/contrib/v3/casbin/v2"
   _ "github.com/go-sql-driver/mysql"
-  "github.com/casbin/xorm-adapter/v2"
+  gormadapter "github.com/casbin/gorm-adapter/v3"
 )
 
 func main() {
   app := fiber.New()
 
+  adapter, _ := gormadapter.NewAdapter("mysql", "root:@tcp(127.0.0.1:3306)/")
   authz := casbin.New(casbin.Config{
       ModelFilePath: "path/to/rbac_model.conf",
-      PolicyAdapter: xormadapter.NewAdapter("mysql", "root:@tcp(127.0.0.1:3306)/"),
+      PolicyAdapter: adapter,
       Lookup: func(c fiber.Ctx) string {
-          // fetch authenticated user subject
+          return "" // fetch authenticated user subject
       },
   })
 
@@ -80,7 +82,7 @@ func main() {
         // your handler
       },
   )
-  
+
   app.Delete("/blog/:id",
     authz.RequiresPermissions([]string{"blog:create", "blog:delete"}, casbin.WithValidationRule(casbin.AtLeastOneRule)),
     func(c fiber.Ctx) error {
@@ -99,19 +101,20 @@ package main
 
 import (
   "github.com/gofiber/fiber/v3"
-  "github.com/gofiber/contrib/v3/casbin"
+  "github.com/gofiber/contrib/v3/casbin/v2"
   _ "github.com/go-sql-driver/mysql"
-  "github.com/casbin/xorm-adapter/v2"
+  gormadapter "github.com/casbin/gorm-adapter/v3"
 )
 
 func main() {
   app := fiber.New()
 
+  adapter, _ := gormadapter.NewAdapter("mysql", "root:@tcp(127.0.0.1:3306)/")
   authz := casbin.New(casbin.Config{
       ModelFilePath: "path/to/rbac_model.conf",
-      PolicyAdapter: xormadapter.NewAdapter("mysql", "root:@tcp(127.0.0.1:3306)/"),
+      PolicyAdapter: adapter,
       Lookup: func(c fiber.Ctx) string {
-          // fetch authenticated user subject
+          return "" // fetch authenticated user subject
       },
   })
 
@@ -134,22 +137,23 @@ package main
 
 import (
   "github.com/gofiber/fiber/v3"
-  "github.com/gofiber/contrib/v3/casbin"
+  "github.com/gofiber/contrib/v3/casbin/v2"
   _ "github.com/go-sql-driver/mysql"
-  "github.com/casbin/xorm-adapter/v2"
+  gormadapter "github.com/casbin/gorm-adapter/v3"
 )
 
 func main() {
   app := fiber.New()
 
+  adapter, _ := gormadapter.NewAdapter("mysql", "root:@tcp(127.0.0.1:3306)/")
   authz := casbin.New(casbin.Config{
       ModelFilePath: "path/to/rbac_model.conf",
-      PolicyAdapter: xormadapter.NewAdapter("mysql", "root:@tcp(127.0.0.1:3306)/"),
+      PolicyAdapter: adapter,
       Lookup: func(c fiber.Ctx) string {
-          // fetch authenticated user subject
+          return "" // fetch authenticated user subject
       },
   })
-  
+
   app.Put("/blog/:id",
     authz.RequiresRoles([]string{"admin"}),
     func(c fiber.Ctx) error {
